Re-completed Labs 1 & 2
The videos for labs 1 and 2 have been followed to speedily return to a position of a finished lab 4. Once lab 4 is complete, own work can be put into the coursework (should there remain time) which will differentiate it from others' projects.
Unfortunately, however, following lab 2 to a tee (according to the video from the lecturer) displays no torus where there should be one. Frustrating as this is, it shall hopefully be remedied further on.