Implemented Multiple Lights
The project now has three lights in it all affecting the colouration and lighting of the teapot. Unfortunately, however, it appears the ambient light is the only one functioning, or at least that the diffuse and specular are making miniscule difference. There is only a flat colour able to be put on the teapot.