Continuing to CW2
Coursework 2 has finally been started. Struggles with university and life alike have proven difficult and have obstructed this coursework's progress.
However, the first part of what is desired is done. The Edge Detection lecture was followed. There are no technical errors, but it seems that, ironically, CW1 was too far ahead and so not all parts of the lecture video fit with the pre-exisitng CW1 code. Hence, a black screen or a grey screen with black streaks. But, it is the Gaussian Blur that is the desired effect, and this continues, lecture-wise, from Edge Detection, hence doing this first despite not wanting it.