Implemented Gaussian Tutorial
The lecture video on Gaussian Blur has been followed. However, next to happen will actually be finishing week 4, and completing Edge Detection and Gaussian Blur line-for-line, as the inconsistencies between the CW code and the example are likely where the errors are. Hence, completing the baseline lecture parts can allow for cross-referencing.