Attempted Fixes
At one point a single colour appeared, but it was uncertain as to why and why the plane did not appear as it had done before. The likelihood was the transition from multiple lights to one.

It seems the best course of action is to continue (again) straight from lab one, but with assistance when problems arise, instead of retroactively.